Enabling component-click for premium learning components
