+++ scripts/browse-card/browse-cards-premium-learning.js
@@ -3,6 +3,7 @@ import { loadCSS, decorateIcons } from '../lib-franklin.js';
 import { fetchLanguagePlaceholders } from '../scripts.js';
 import UserActions from '../user-actions/user-actions.js';
 import { pushBrowseCardClickEvent } from '../analytics/lib-analytics.js';
+import { getCardHeaderAndPosition } from './browse-card-utils.js';
 
 /**
  * @fileoverview premium-learning specific browse card implementation
@@ -16,36 +17,6 @@ try {
 } catch (err) {
   // eslint-disable-next-line no-console
   console.error('Error fetching placeholders:', err);
-}
-
-/**
- * Calculates card header and position for analytics tracking
- * @param {HTMLElement} card - The card element
- * @param {HTMLElement} element - The container element
- * @returns {Object} Object with cardHeader and cardPosition
- * @private
- */
-function getCardHeaderAndPosition(card, element) {
-  let cardHeader = '';
-  const currentBlock = card.closest('.block');
-  const headerEl = currentBlock?.querySelector(
-    '.browse-cards-block-title, .rec-block-header, .inprogress-courses-header-wrapper',
-  );
-  if (headerEl) {
-    const cloned = headerEl.cloneNode(true);
-    cloned.querySelectorAll('[data-cs-mask]').forEach((el) => el.remove());
-    cardHeader = cloned.innerText.trim();
-  }
-
-  cardHeader = cardHeader || currentBlock?.getAttribute('data-block-name')?.trim() || '';
-
-  let cardPosition = '';
-  if (element?.parentElement?.children) {
-    const siblings = Array.from(element.parentElement.children);
-    cardPosition = String(siblings.indexOf(element) + 1);
-  }
-
-  return { cardHeader, cardPosition };
 }
 
 /**
@@ -298,11 +269,14 @@ export async function buildPLCard(element, model) {
     const cardContainer = document.createElement('a');
     cardContainer.href = model.viewLink;
 
+    if (element.closest('.section')?.getAttribute('data-new-tab') === 'true') {
+      cardContainer.setAttribute('target', '_blank');
+      cardContainer.setAttribute('rel', 'noopener noreferrer');
+    }
+
     // Prevent navigation when clicking user actions
     cardContainer.addEventListener('click', (e) => {
-      const preventLinkRedirection = !!(e.target && e.target.closest('.user-actions'));
-
-      if (preventLinkRedirection) {
+      if (e.target?.closest('.user-actions')) {
         e.preventDefault();
       }
     });
@@ -315,11 +289,6 @@ export async function buildPLCard(element, model) {
 
   element.querySelector('a')?.addEventListener('click', (e) => {
     const { cardHeader, cardPosition } = getCardHeaderAndPosition(card, element);
-    const shouldOpenInNewTab = element.closest('.section')?.getAttribute('data-new-tab') === 'true';
-
-    if (shouldOpenInNewTab) {
-      element.querySelector('a')?.setAttribute('target', '_blank');
-    }
 
     const cardActions = card.querySelector('.premium-learning-card-actions');
     if (cardActions) {
@@ -331,9 +300,7 @@ export async function buildPLCard(element, model) {
       return;
     }
 
-    if (e.target.closest('a:not(.user-actions)')) {
-      pushBrowseCardClickEvent('browseCardClicked', model, cardHeader, cardPosition);
-    }
+    pushBrowseCardClickEvent('browseCardClicked', model, cardHeader, cardPosition);
   });
 }
